requirements.txt
fastapi
uvicorn
pydantic-settings
asyncpg
redis
pydantic
python-dotenv
pwdlib[argon2]
PyJWT
email-validator
python-multipart